requirements.txt
openai==0.27.7
Pygments==2.15.1
python-dotenv==1.0.0
python-telegram-bot==20.3
Requests==2.31.0
SQLAlchemy==2.0.15
tiktoken==0.4.0
